Task 4.1 add groups CRUD api

+++ src/controllers/user.controller.ts
@@ -40,7 +40,7 @@ export const create = async (req: ValidatedRequest<UserRequestSchema>, res: Resp
   try {
     const user = req.body;
 
-    const newUser = await UserService.create(user as User);
+    const newUser = await UserService.create(user as BaseUser);
 
     res.status(CREATED).send(newUser);
   } catch (e) {
@@ -69,9 +69,13 @@ export const update = async (req: ValidatedRequest<UserRequestSchema>, res: Resp
 export const remove = async (req: Request, res: Response) => {
   try {
     const id: string = req.params.id;
-    await UserService.remove(id);
+    const existingUser = await UserService.findOne(id);
 
-    res.sendStatus(NO_CONTENT);
+    if (existingUser) {
+      await UserService.remove(id);
+      return res.sendStatus(NO_CONTENT);
+    }
+    return res.status(NOT_FOUND).send('User not found');
   } catch (e) {
     res.status(INTERNAL_SERVER_ERROR).send((e as Error).message);
   }
